% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eksi_date_time_parser.R
\name{eksi_date_time_parse}
\alias{eksi_date_time_parse}
\title{Convert data_time column into POSIXct}
\usage{
eksi_date_time_parse(x)
}
\arguments{
\item{x}{A tibble. The result of scrape function will be the input of this function.}
}
\value{
A tibble. Adds two new columns (POSIXct type) to the existing tibble called original_date_time and edited_date_time. The latter is only for edited entries. Otherwise, fills NA.
}
\description{
Convert data_time column into POSIXct
}
\examples{
scrape_entry(3000010) \%>\% eksi_date_time_parse()
}
