% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eksi_originate_topic_url}
\alias{eksi_originate_topic_url}
\title{Originate topic urls from eksisozluk}
\usage{
eksi_originate_topic_url(x)
}
\arguments{
\item{x}{Character. Takes eksisozluk topic url copied from browser.}
}
\value{
This small function stips the url from page numbers, dates, focus to certain entry IDs etc which all come after '?'
}
\description{
Originate topic urls from eksisozluk
}
