% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eksiR-package.r
\docType{package}
\name{eksiR}
\alias{eksiR}
\alias{eksiR-package}
\title{Scrape Entries From eksisozluk.com}
\description{
A webscraper for eksisozluk.com (sour dictionary) which is
one of the top 20 sites in Turkey (https://www.alexa.com/topsites/countries/TR).
It allows you to scrape data by entry number and returns a neat tibble. It
also returns informative errors for deleted entries.
}
\details{
eksiR.
}
\author{
person("Sefa", "Ozalp", email = "ozalpas [at] cardiff.ac.uk", role = c("aut", "cre"))
}
